requirements.txt
django
djangorestframework
scikit-learn
pandas
matplotlib
joblib
pillow